requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
django-browser-reload==1.18.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==10.4.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
qrcode==8.0
setuptools==69.2.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.43.0
whitenoise==6.11.0
